修改: 檔案結構與各元件路徑、prettier 保留分號、App.vue 入口路徑
