+++ src/App.vue
@@ -1,13 +1,21 @@
 <script setup lang="ts">
-import HeaderNavbar from './components/HeaderNavbar.vue';
-import Home from './components/Home.vue'
-import Footer from './components/Footer.vue';
+import HeaderNavbar from './components/layout/HeaderNavbar.vue';
+import Home from './views/Home/Home.vue';
+import Footer from './components/layout/Footer.vue';
+import SearchResults from './views/Search/SearchResults.vue';
+import DetailView from './views/Search/components/DetailView.vue';
+import ServiceSearchFlow from './components/service-search/ServiceSearchFlow.vue';
+import VehicleManagement from './views/User/VehicleManagement.vue';
 </script>
 
 <template>
   <HeaderNavbar />
   <Home />
   <Footer />
+  <SearchResults />
+  <DetailView />
+  <ServiceSearchFlow />
+  <VehicleManagement />
 </template>
 
 <style scoped></style>
