+++ src/components/service-search/ServiceSearchFlow.vue
@@ -1,30 +1,30 @@
 <script setup lang="ts">
-import { computed, ref } from 'vue'
-import ServiceStepper from './ServiceStepper.vue'
-import ServiceModal from './ServiceModal.vue'
-import BookingTimeSelectorView from '../BookingTimeSelectorView.vue'
+import { computed, ref } from 'vue';
+import ServiceStepper from './ServiceStepper.vue';
+import ServiceModal from './ServiceModal.vue';
+import BookingTimeSelectorView from '../../views/Search/components/BookingTimeSelectorView.vue';
 
 defineEmits<{
-  (e: 'back'): void
-}>()
+  (e: 'back'): void;
+}>();
 
-type StepKey = 'service' | 'datetime' | 'contact' | 'confirm'
+type StepKey = 'service' | 'datetime' | 'contact' | 'confirm';
 const steps: Array<{ key: StepKey; label: string }> = [
   { key: 'service', label: '選擇服務' },
   { key: 'datetime', label: '日期時段' },
   { key: 'contact', label: '聯絡資訊' },
   { key: 'confirm', label: '確認預約' },
-]
+];
 
-const step = ref<number>(1)
+const step = ref<number>(1);
 
-type ServiceKey = 'oil' | 'tire' | 'brake' | 'maintain'
+type ServiceKey = 'oil' | 'tire' | 'brake' | 'maintain';
 const serviceOptions: Array<{
-  key: ServiceKey
-  title: string
-  minutes: number
-  fa: string
-  icon: string
+  key: ServiceKey;
+  title: string;
+  minutes: number;
+  fa: string;
+  icon: string;
 }> = [
   { key: 'oil', title: '機油更換', minutes: 30, fa: 'fa-solid fa-oil-can', icon: 'gas_meter' },
   {
@@ -48,46 +48,46 @@ const serviceOptions: Array<{
     fa: 'fa-solid fa-screwdriver-wrench',
     icon: 'construction',
   },
-]
-const selectedService = ref<ServiceKey | null>(null)
-const selectedDate = ref<Date | null>(null)
-const selectedTime = ref<string | null>(null)
-const contactName = ref<string>('')
-const contactPhone = ref<string>('')
-const contactEmail = ref<string>('')
-const contactNote = ref<string>('')
+];
+const selectedService = ref<ServiceKey | null>(null);
+const selectedDate = ref<Date | null>(null);
+const selectedTime = ref<string | null>(null);
+const contactName = ref<string>('');
+const contactPhone = ref<string>('');
+const contactEmail = ref<string>('');
+const contactNote = ref<string>('');
 
 const isContactValid = computed(() => {
-  return !!contactName.value.trim() && !!contactPhone.value.trim() && !!contactEmail.value.trim()
-})
+  return !!contactName.value.trim() && !!contactPhone.value.trim() && !!contactEmail.value.trim();
+});
 
 const maxUnlocked = computed(() => {
-  if (step.value === 1) return selectedService.value ? 2 : 1
-  if (step.value === 2) return selectedDate.value && selectedTime.value ? 3 : 2
-  if (step.value === 3) return isContactValid.value ? 4 : 3
-  return 4
-})
+  if (step.value === 1) return selectedService.value ? 2 : 1;
+  if (step.value === 2) return selectedDate.value && selectedTime.value ? 3 : 2;
+  if (step.value === 3) return isContactValid.value ? 4 : 3;
+  return 4;
+});
 
 const goToStep = (n: number) => {
-  const target = Math.min(4, Math.max(1, n))
-  step.value = Math.min(target, maxUnlocked.value)
-}
+  const target = Math.min(4, Math.max(1, n));
+  step.value = Math.min(target, maxUnlocked.value);
+};
 
 const nextStep = () => {
-  if (!canGoNext.value) return
-  step.value = Math.min(4, step.value + 1)
-}
+  if (!canGoNext.value) return;
+  step.value = Math.min(4, step.value + 1);
+};
 
 const prevStep = () => {
-  step.value = Math.max(1, step.value - 1)
-}
+  step.value = Math.max(1, step.value - 1);
+};
 
 const canGoNext = computed(() => {
-  if (step.value === 1) return selectedService.value !== null
-  if (step.value === 2) return selectedDate.value !== null && selectedTime.value !== null
-  if (step.value === 3) return isContactValid.value
-  return true
-})
+  if (step.value === 1) return selectedService.value !== null;
+  if (step.value === 2) return selectedDate.value !== null && selectedTime.value !== null;
+  if (step.value === 3) return isContactValid.value;
+  return true;
+});
 
 const canFinish = computed(() => {
   return (
@@ -95,50 +95,50 @@ const canFinish = computed(() => {
     selectedDate.value !== null &&
     selectedTime.value !== null &&
     isContactValid.value
-  )
-})
+  );
+});
 
 const serviceCardClass = (key: ServiceKey) => {
   return selectedService.value === key
     ? 'border-[#6B6B5C] bg-[#6B6B5C] shadow-[0_3px_0_rgba(0,0,0,0.18)]'
-    : 'border-[#E6E6DF] bg-white hover:border-[#CFCFC6]'
-}
+    : 'border-[#E6E6DF] bg-white hover:border-[#CFCFC6]';
+};
 
 const prevBtnClass = computed(() => {
   return step.value === 1
     ? 'border-[#E6E6DF] text-[#B5B5AD] bg-[#FBFAF7] cursor-not-allowed'
-    : 'border-[#E6E6DF] text-[#6B6B5C] bg-[#FBFAF7] hover:bg-white'
-})
+    : 'border-[#E6E6DF] text-[#6B6B5C] bg-[#FBFAF7] hover:bg-white';
+});
 const nextBtnClass = computed(() => {
   return canGoNext.value
     ? 'bg-[#6B6B5C] hover:opacity-95 active:opacity-90 shadow-[0_3px_0_rgba(0,0,0,0.18)]'
-    : 'bg-[#D1D1CB] cursor-not-allowed'
-})
+    : 'bg-[#D1D1CB] cursor-not-allowed';
+});
 
 const selectedDateLong = computed(() => {
-  if (!selectedDate.value) return ''
+  if (!selectedDate.value) return '';
   return selectedDate.value.toLocaleDateString('zh-TW', {
     year: 'numeric',
     month: 'long',
     day: 'numeric',
     weekday: 'long',
-  })
-})
+  });
+});
 
-const selectedTimeLong = computed(() => selectedTime.value || '')
+const selectedTimeLong = computed(() => selectedTime.value || '');
 const selectedServiceObj = computed(
   () => serviceOptions.find((s) => s.key === selectedService.value) || null,
-)
+);
 
 interface SummaryItem {
-  key: string
-  icon: string
-  iconClass?: string
-  label: string
-  value: string
-  valueClass?: string
-  details?: string
-  detailsClass?: string
+  key: string;
+  icon: string;
+  iconClass?: string;
+  label: string;
+  value: string;
+  valueClass?: string;
+  details?: string;
+  detailsClass?: string;
 }
 
 const summaryItems = computed<SummaryItem[]>(() => [
@@ -189,81 +189,58 @@ const summaryItems = computed<SummaryItem[]>(() => [
     value: contactNote.value || '—',
     valueClass: 'whitespace-pre-wrap',
   },
-])
+]);
 
-const showSuccess = ref(false)
-const bookingCode = ref<string>('')
+const showSuccess = ref(false);
+const bookingCode = ref<string>('');
 
 const randomBookingCode = () => {
-  const n = Math.floor(10000000 + Math.random() * 90000000)
-  return `BK${n}`
-}
+  const n = Math.floor(10000000 + Math.random() * 90000000);
+  return `BK${n}`;
+};
 
 const finish = () => {
-  if (!canFinish.value) return
-  bookingCode.value = randomBookingCode()
-  showSuccess.value = true
-}
+  if (!canFinish.value) return;
+  bookingCode.value = randomBookingCode();
+  showSuccess.value = true;
+};
 
 const resetAll = () => {
-  showSuccess.value = false
-  step.value = 1
-  selectedService.value = null
-  selectedDate.value = null
-  selectedTime.value = null
-  contactName.value = ''
-  contactPhone.value = ''
-  contactEmail.value = ''
-  contactNote.value = ''
-}
+  showSuccess.value = false;
+  step.value = 1;
+  selectedService.value = null;
+  selectedDate.value = null;
+  selectedTime.value = null;
+  contactName.value = '';
+  contactPhone.value = '';
+  contactEmail.value = '';
+  contactNote.value = '';
+};
 </script>
 
 <template>
   <div class="min-h-screen bg-[#FAF8F5] px-6 py-10 text-[#2B2B2B]">
     <div class="mx-auto mb-6 w-full max-w-6xl">
-      <button
-        type="button"
-        class="flex items-center gap-2 rounded-xl border border-[#E6E6DF] bg-white px-6 py-3 text-sm font-semibold text-[#6B6B5C] transition hover:bg-[#F5F4EF]"
-        @click="$emit('back')"
-      >
+      <button type="button" class="flex items-center gap-2 rounded-xl border border-[#E6E6DF] bg-white px-6 py-3 text-sm font-semibold text-[#6B6B5C] transition hover:bg-[#F5F4EF]" @click="$emit('back')">
         <!-- <i class="fa-solid fa-arrow-left"></i> -->
         <span class="material-symbols-outlined"> arrow_left_alt </span>
         返回維修廠介紹頁面
       </button>
     </div>
-    <div
-      class="mx-auto w-full max-w-6xl rounded-[28px] border border-[#E6E6DF] bg-white shadow-[0_2px_10px_rgba(0,0,0,0.06)]"
-    >
+    <div class="mx-auto w-full max-w-6xl rounded-[28px] border border-[#E6E6DF] bg-white shadow-[0_2px_10px_rgba(0,0,0,0.06)]">
       <div class="p-10">
         <ServiceStepper :step="step" :steps="steps" @go="goToStep" />
         <div class="mt-16">
           <div v-if="step === 1">
             <h2 class="text-3xl font-bold tracking-tight">選擇服務項目</h2>
             <div class="mt-10 grid grid-cols-2 gap-6">
-              <button
-                v-for="item in serviceOptions"
-                :key="item.key"
-                type="button"
-                class="flex items-center gap-5 rounded-2xl border px-7 py-6 text-left transition"
-                :class="serviceCardClass(item.key)"
-                @click="selectedService = item.key"
-              >
-                <div
-                  class="grid h-14 w-14 place-items-center rounded-2xl transition"
-                  :class="
-                    selectedService === item.key
-                      ? 'bg-white/20 text-white'
-                      : 'bg-[#F5F4EF] text-[#6B6B5C]'
-                  "
-                >
+              <button v-for="item in serviceOptions" :key="item.key" type="button" class="flex items-center gap-5 rounded-2xl border px-7 py-6 text-left transition" :class="serviceCardClass(item.key)" @click="selectedService = item.key">
+                <div class="grid h-14 w-14 place-items-center rounded-2xl transition" :class=" selectedService === item.key ? 'bg-white/20 text-white' : 'bg-[#F5F4EF] text-[#6B6B5C]'">
                   <!-- <i :class="item.fa" class="text-xl" aria-hidden="true"></i> -->
                   <span class="material-symbols-outlined">{{ item.icon }}</span>
                 </div>
                 <div class="min-w-0">
-                  <div
-                    class="text-lg font-semibold transition"
-                    :class="selectedService === item.key ? 'text-white' : 'text-[#2B2B2B]'"
-                  >
+                  <div class="text-lg font-semibold transition" :class="selectedService === item.key ? 'text-white' : 'text-[#2B2B2B]'">
                     {{ item.title }}
                   </div>
                 </div>
@@ -283,11 +260,7 @@ const resetAll = () => {
                 </div>
                 <div>
                   <div class="text-base font-semibold text-[#6B6B5C]">姓名</div>
-                  <input
-                    v-model.trim="contactName"
-                    class="mt-3 h-11 w-full rounded-xl bg-[#F5F4EF] px-4 text-sm outline-none placeholder:text-[#B5B5AD]"
-                    placeholder="請輸入姓名"
-                  />
+                  <input v-model.trim="contactName" class="mt-3 h-11 w-full rounded-xl bg-[#F5F4EF] px-4 text-sm outline-none placeholder:text-[#B5B5AD]" placeholder="請輸入姓名"/>
                 </div>
               </div>
               <div class="h-px bg-[#E6E6DF]" />
@@ -298,12 +271,7 @@ const resetAll = () => {
                 </div>
                 <div>
                   <div class="text-base font-semibold text-[#6B6B5C]">電話</div>
-                  <input
-                    v-model.trim="contactPhone"
-                    inputmode="tel"
-                    class="mt-3 h-11 w-full rounded-xl bg-[#F5F4EF] px-4 text-sm outline-none placeholder:text-[#B5B5AD]"
-                    placeholder="請輸入電話"
-                  />
+                  <input v-model.trim="contactPhone" inputmode="tel" class="mt-3 h-11 w-full rounded-xl bg-[#F5F4EF] px-4 text-sm outline-none placeholder:text-[#B5B5AD]" placeholder="請輸入電話"/>
                 </div>
               </div>
               <div class="h-px bg-[#E6E6DF]" />
@@ -314,12 +282,7 @@ const resetAll = () => {
                 </div>
                 <div>
                   <div class="text-base font-semibold text-[#6B6B5C]">電子郵件</div>
-                  <input
-                    v-model.trim="contactEmail"
-                    inputmode="email"
-                    class="mt-3 h-11 w-full rounded-xl bg-[#F5F4EF] px-4 text-sm outline-none placeholder:text-[#B5B5AD]"
-                    placeholder="請輸入電子郵件"
-                  />
+                  <input v-model.trim="contactEmail" inputmode="email" class="mt-3 h-11 w-full rounded-xl bg-[#F5F4EF] px-4 text-sm outline-none placeholder:text-[#B5B5AD]" placeholder="請輸入電子郵件"/>
                 </div>
               </div>
               <div class="h-px bg-[#E6E6DF]" />
@@ -330,12 +293,7 @@ const resetAll = () => {
                 </div>
                 <div>
                   <div class="text-base font-semibold text-[#6B6B5C]">備註</div>
-                  <textarea
-                    v-model.trim="contactNote"
-                    rows="3"
-                    class="mt-3 w-full resize-none rounded-xl bg-[#F5F4EF] px-4 py-3 text-sm outline-none placeholder:text-[#B5B5AD]"
-                    placeholder="如有特殊需求請在此填寫"
-                  />
+                  <textarea v-model.trim="contactNote" rows="3" class="mt-3 w-full resize-none rounded-xl bg-[#F5F4EF] px-4 py-3 text-sm outline-none placeholder:text-[#B5B5AD]" placeholder="如有特殊需求請在此填寫"/>
                 </div>
               </div>
             </div>
@@ -347,20 +305,14 @@ const resetAll = () => {
                 <div class="grid grid-cols-[44px_1fr] gap-6 py-8">
                   <div class="pt-1" :class="item.iconClass || 'text-[#6B6B5C]'">
                     <!-- <i :class="[item.icon, 'text-2xl']" aria-hidden="true"></i> -->
-                    <span class="material-symbols-outlined" aria-hidden="true">{{
-                      item.icon
-                    }}</span>
+                    <span class="material-symbols-outlined" aria-hidden="true">{{ item.icon }}</span>
                   </div>
                   <div>
                     <div class="text-base font-semibold text-[#6B6B5C]">{{ item.label }}</div>
                     <div class="mt-2 text-lg font-semibold" :class="item.valueClass">
                       {{ item.value }}
                     </div>
-                    <div
-                      v-if="item.details"
-                      class="mt-2"
-                      :class="item.detailsClass || 'text-sm text-[#7A7A7A]'"
-                    >
+                    <div v-if="item.details" class="mt-2" :class="item.detailsClass || 'text-sm text-[#7A7A7A]'">
                       {{ item.details }}
                     </div>
                   </div>
@@ -371,42 +323,18 @@ const resetAll = () => {
           </div>
         </div>
         <div class="mt-14 flex items-center justify-between">
-          <button
-            type="button"
-            class="rounded-xl border px-10 py-3 text-sm font-semibold transition"
-            :class="prevBtnClass"
-            :disabled="step === 1"
-            @click="prevStep"
-          >
+          <button type="button" class="rounded-xl border px-10 py-3 text-sm font-semibold transition" :class="prevBtnClass" :disabled="step === 1" @click="prevStep">
             上一步
           </button>
-          <button
-            v-if="step < 4"
-            type="button"
-            class="rounded-xl px-10 py-3 text-sm font-semibold text-white transition"
-            :class="nextBtnClass"
-            :disabled="!canGoNext"
-            @click="nextStep"
-          >
+          <button v-if="step < 4" type="button" class="rounded-xl px-10 py-3 text-sm font-semibold text-white transition" :class="nextBtnClass" :disabled="!canGoNext" @click="nextStep">
             下一步
           </button>
-          <button
-            v-else
-            type="button"
-            class="rounded-xl bg-[#6B6B5C] px-10 py-3 text-sm font-semibold text-white shadow-[0_3px_0_rgba(0,0,0,0.18)] hover:opacity-95 active:opacity-90"
-            :disabled="!canFinish"
-            @click="finish"
-          >
+          <button v-else type="button" class="rounded-xl bg-[#6B6B5C] px-10 py-3 text-sm font-semibold text-white shadow-[0_3px_0_rgba(0,0,0,0.18)] hover:opacity-95 active:opacity-90" :disabled="!canFinish" @click="finish">
             完成預約
           </button>
         </div>
       </div>
     </div>
-    <ServiceModal
-      v-if="showSuccess"
-      :bookingCode="bookingCode"
-      @close="showSuccess = false"
-      @again="resetAll"
-    />
+    <ServiceModal v-if="showSuccess" :bookingCode="bookingCode" @close="showSuccess = false" @again="resetAll"/>
   </div>
 </template>
