+++ src/components/service-search/ServiceStepper.vue
@@ -1,14 +1,14 @@
 <script setup lang="ts">
-import { computed } from "vue";
+import { computed } from 'vue';
 
-type StepKey = "service" | "datetime" | "contact" | "confirm";
+type StepKey = 'service' | 'datetime' | 'contact' | 'confirm';
 
 const props = defineProps<{
   step: number;
   steps: Array<{ key: StepKey; label: string }>;
 }>();
 
-defineEmits<{ (e: "go", n: number): void }>();
+defineEmits<{ (e: 'go', n: number): void }>();
 
 const progressWidth = computed(() => {
   const segments = props.steps.length - 1;
@@ -17,24 +17,20 @@ const progressWidth = computed(() => {
 });
 
 const circleClass = (n: number) => {
-  if (n <= props.step) return "bg-[#6B6B5C] text-white";
-  return "bg-white text-[#7A7A7A] border border-[#E6E6DF]";
+  if (n <= props.step) return 'bg-[#6B6B5C] text-white';
+  return 'bg-white text-[#7A7A7A] border border-[#E6E6DF]';
 };
 
 const labelClass = (n: number) => {
-  if (n === props.step) return "text-[#2B2B2B] font-semibold";
-  return "text-[#7A7A7A]";
+  if (n === props.step) return 'text-[#2B2B2B] font-semibold';
+  return 'text-[#7A7A7A]';
 };
 </script>
 
 <template>
   <div class="relative w-full">
     <div class="absolute left-0 right-0 top-[18px] h-px bg-[#E6E6DF]" />
-    <div
-      v-if="step > 1"
-      class="absolute left-0 top-[18px] h-px bg-[#6B6B5C] transition-all duration-300"
-      :style="{ width: progressWidth }"
-    />
+    <div v-if="step > 1" class="absolute left-0 top-[18px] h-px bg-[#6B6B5C] transition-all duration-300" :style="{ width: progressWidth }"/>
     <div class="grid grid-cols-4 items-start gap-0">
       <button
         v-for="(s, idx) in steps"
@@ -43,10 +39,7 @@ const labelClass = (n: number) => {
         class="group relative flex flex-col items-center gap-3"
         @click="$emit('go', idx + 1)"
       >
-        <span 
-          class="grid h-9 w-9 place-items-center rounded-full text-sm font-semibold transition z-10" 
-          :class="circleClass(idx + 1)"
-        >
+        <span class="grid h-9 w-9 place-items-center rounded-full text-sm font-semibold transition z-10" :class="circleClass(idx + 1)">
           <template v-if="idx + 1 < step">
             <i class="fa-solid fa-check text-[14px]" aria-hidden="true"></i>
           </template>
