+++ src/views/Auth/LoginMode.vue
@@ -44,12 +44,12 @@ const handleLoginSubmit = () => {
         type="button"
         aria-label="關閉"
       >
-        <svg 
-          xmlns="http://www.w3.org/2000/svg" 
-          fill="none" 
-          viewBox="0 0 24 24" 
-          stroke-width="2.5" 
-          stroke="currentColor" 
+        <svg
+          xmlns="http://www.w3.org/2000/svg"
+          fill="none"
+          viewBox="0 0 24 24"
+          stroke-width="2.5"
+          stroke="currentColor"
           class="w-5 h-5"
         >
           <path stroke-linecap="round" stroke-linejoin="round" d="M6 18L18 6M6 6l12 12" />
@@ -64,9 +64,16 @@ const handleLoginSubmit = () => {
             v-model="loginForm.account"
             type="text"
             placeholder="請輸入您的帳號"
-            :class="['w-full px-4 py-2 text-[16px] rounded-lg border outline-none transition-all', formErrors.account ? 'border-red-500 bg-red-50' : 'border-gray-300 focus:border-[#6b6b5a]']"
+            :class="[
+              'w-full px-4 py-2 text-[16px] rounded-lg border outline-none transition-all',
+              formErrors.account
+                ? 'border-red-500 bg-red-50'
+                : 'border-gray-300 focus:border-[#6b6b5a]',
+            ]"
           />
-          <span v-if="formErrors.account" class="text-[14px] text-red-500">{{ formErrors.account }}</span>
+          <span v-if="formErrors.account" class="text-[14px] text-red-500">
+            {{ formErrors.account }}
+          </span>
         </div>
         <div class="flex flex-col gap-1">
           <label for="password" class="text-[16px] text-gray-700 font-medium">密碼</label>
@@ -75,25 +82,50 @@ const handleLoginSubmit = () => {
             v-model="loginForm.password"
             type="password"
             placeholder="請輸入您的密碼"
-            :class="['w-full px-4 py-2 text-[16px] rounded-lg border outline-none transition-all', formErrors.password ? 'border-red-500 bg-red-50' : 'border-gray-300 focus:border-[#6b6b5a]']"
+            :class="[
+              'w-full px-4 py-2 text-[16px] rounded-lg border outline-none transition-all',
+              formErrors.password
+                ? 'border-red-500 bg-red-50'
+                : 'border-gray-300 focus:border-[#6b6b5a]',
+            ]"
           />
-          <span v-if="formErrors.password" class="text-[14px] text-red-500">{{ formErrors.password }}</span>
+          <span v-if="formErrors.password" class="text-[14px] text-red-500">
+            {{ formErrors.password}}
+          </span>
         </div>
         <div class="flex justify-between items-center mt-1">
           <label class="flex items-center gap-2 cursor-pointer select-none">
-            <input v-model="loginForm.rememberMe" type="checkbox" class="w-4 h-4 text-[#6b6b5a] rounded border-gray-300 focus:ring-[#6b6b5a]" />
+            <input
+              v-model="loginForm.rememberMe"
+              type="checkbox"
+              class="w-4 h-4 text-[#6b6b5a] rounded border-gray-300 focus:ring-[#6b6b5a]"
+            />
             <span class="text-[14px] text-gray-600">記住帳號密碼</span>
           </label>
-          <a href="#" class="text-[14px] text-[#6b6b5a] hover:underline hover:text-[#4a4a43]">忘記密碼？</a>
+          <a href="#" class="text-[14px] text-[#6b6b5a] hover:underline hover:text-[#4a4a43]">
+            忘記密碼？
+          </a>
         </div>
-        <button type="submit" class="mt-2 w-full py-2.5 text-[16px] font-bold text-white bg-[#6b6b5a] rounded-lg hover:bg-[#5a5a4a] transition-colors shadow-md cursor-pointer">登入</button>
+        <button
+          type="submit"
+          class="mt-2 w-full py-2.5 text-[16px] font-bold text-white bg-[#6b6b5a] rounded-lg hover:bg-[#5a5a4a] transition-colors shadow-md cursor-pointer"
+        >
+          登入
+        </button>
       </form>
       <div class="relative flex justify-center items-center my-6">
-        <div class="absolute inset-0 flex items-center"><div class="w-full border-t border-gray-300"></div></div>
-        <div class="relative px-4 text-[14px] text-gray-500 bg-white/0 backdrop-blur-md">還沒有帳號嗎？</div>
+        <div class="absolute inset-0 flex items-center">
+          <div class="w-full border-t border-gray-300"></div>
+        </div>
+        <div class="relative px-4 text-[14px] text-gray-500 bg-white/0 backdrop-blur-md">
+          還沒有帳號嗎？
+        </div>
       </div>
       <div class="text-center">
-        <button @click="$emit('switch-to-signup')" class="text-[16px] font-bold text-[#6b6b5a] hover:underline cursor-pointer">
+        <button
+          @click="$emit('switch-to-signup')"
+          class="text-[16px] font-bold text-[#6b6b5a] hover:underline cursor-pointer"
+        >
           立即註冊
         </button>
       </div>
