+++ src/views/Search/components/DetailView.vue
